Update synchronize_event_handler.py: 操作本地文件之前先检查是否存在
(cherry picked from commit 7c4d3390598b15ba70d769a30b4fac905bfeca3a)

+++ cloudsync/synchronize_event_handler.py
@@ -55,6 +55,12 @@ class SynchronizeEventHandler:
         from_path = self.from_path if from_path is None else from_path
         to_path = self.to_path if to_path is None else to_path
         logger.info('准备将云端文件夹 {from_path} 下载到本地文件夹 {to_path}'.format(from_path=from_path, to_path=to_path))
+
+        # check if the local folder exits
+        if os.path.exists(to_path):
+            return
+
+        # create folder
         os.mkdir(to_path)
         logger.info('已创建本地文件夹 {to_path}'.format(to_path=to_path))
         logger.info('遍历云端文件夹 {from_path} 中的内容，递归下载'.format(from_path=from_path))
@@ -88,6 +94,26 @@ class SynchronizeEventHandler:
         self.cfs.upload(to_path, from_path)
         logger.info('上传本地文件 {from_path} 到云端文件 {to_path} 完成'.format(from_path=from_path, to_path=to_path))
 
+    def download(self, from_path=None, to_path=None):
+        """
+        :param from_path:
+        :param to_path:
+        :return: None
+        """
+        logger = logging.getLogger('{class_name} -> {function_name}'
+                                   .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
+        from_path = self.from_path if from_path is None else from_path
+        to_path = self.to_path if to_path is None else to_path
+
+        # check if the local file exits
+        if os.path.exists(to_path):
+            return
+
+        # download file
+        logger.info('准备将云端文件 {from_path} 下载到本地文件 {to_path}'.format(from_path=from_path, to_path=to_path))
+        self.cfs.download(from_path, to_path)
+        logger.info('下载云端文件 {from_path} 到本地文件 {to_path} 完成'.format(from_path=from_path, to_path=to_path))
+
     def delete_cloud_file(self):
         """
         删除云端文件
@@ -114,6 +140,12 @@ class SynchronizeEventHandler:
         logger = logging.getLogger('{class_name} -> {function_name}'
                                    .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
         local_path = self.from_path
+
+        # check if the local file exits
+        if not os.path.exists(local_path):
+            return
+
+        # delete file
         logger.info('准备删除本地文件 {local_path}'.format(local_path=local_path))
         os.remove(local_path)
         logger.info('删除本地文件 {local_path} 完成'.format(local_path=local_path))
@@ -154,6 +186,12 @@ class SynchronizeEventHandler:
         logger = logging.getLogger('{class_name} -> {function_name}'
                                    .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
         local_path = self.from_path if local_path is None else local_path
+
+        # check if the local folder exits
+        if not os.path.exists(local_path):
+            return
+
+        # delete folder
         logger.info('准备删除本地文件夹 {local_path}'.format(local_path=local_path))
         logger.info('遍历本地文件夹 {local_path} 中的内容，递归删除'.format(local_path=local_path))
         for filename in os.listdir(local_path):
@@ -182,28 +220,12 @@ class SynchronizeEventHandler:
 
     def update_local_file(self):
         """
-        更新本地文件 == 下载文件 hack
         :return: None
         """
         logger = logging.getLogger('{class_name} -> {function_name}'
                                    .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
         from_path = self.from_path
         to_path = self.to_path
-        logger.info('准备将云端文件 {from_path} 下载到本地文件 {to_path}'.format(from_path=from_path, to_path=to_path))
-        self.cfs.download(from_path, to_path)
-        logger.info('下载云端文件 {from_path} 到本地文件 {to_path} 完成'.format(from_path=from_path, to_path=to_path))
-
-    def download(self, from_path=None, to_path=None):
-        """
-        下载文件 == 更新本地文件 hack
-        :param from_path:
-        :param to_path:
-        :return: None
-        """
-        logger = logging.getLogger('{class_name} -> {function_name}'
-                                   .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
-        from_path = self.from_path if from_path is None else from_path
-        to_path = self.to_path if to_path is None else to_path
         logger.info('准备将云端文件 {from_path} 下载到本地文件 {to_path}'.format(from_path=from_path, to_path=to_path))
         self.cfs.download(from_path, to_path)
         logger.info('下载云端文件 {from_path} 到本地文件 {to_path} 完成'.format(from_path=from_path, to_path=to_path))
@@ -236,6 +258,12 @@ class SynchronizeEventHandler:
                                    .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
         from_path = self.from_path
         to_path = self.to_path
+
+        # check if the local file exits
+        if not os.path.exists(from_path):
+            return
+
+        # rename file
         logger.info('准备将本地文件 {from_path} 重命名为 {to_path}'.format(from_path=from_path, to_path=to_path))
         os.rename(from_path, to_path)
         logger.info('重命名本地文件 {from_path} 为 {to_path} 成功'.format(from_path=from_path, to_path=to_path))
@@ -268,6 +296,12 @@ class SynchronizeEventHandler:
                                    .format(class_name=__class__.__name__, function_name=inspect.stack()[0].function))
         from_path = self.from_path
         to_path = self.to_path
+
+        # check if the local folder exits
+        if not os.path.exists(from_path):
+            return
+
+        # rename folder
         logger.info('准备将本地文件夹 {from_path} 重命名为 {to_path}'.format(from_path=from_path, to_path=to_path))
         os.rename(from_path, to_path)
         logger.info('重命名本地文件夹 {from_path} 为 {to_path} 成功'.format(from_path=from_path, to_path=to_path))
